idleness: put tests into a separate module
