+++ src/external/idleness/mod.rs
@@ -2,10 +2,8 @@ mod interface;
 
 pub use interface::*;
 
-pub mod x11;
-#[cfg(test)]
-mod x11_test;
-
 pub mod mock;
+pub mod x11;
+
 #[cfg(test)]
-mod mock_test;
+mod test;
