+++ src/external/idleness/test/mod.rs
@@ -1,0 +1,2 @@
+mod mock_test;
+mod x11_test;
